+++ mod/primitives/pkg/encoding/ssz/merkle/index.go
@@ -23,74 +23,72 @@ package merkle
 import (
 	"slices"
 
-	"github.com/berachain/beacon-kit/mod/errors"
-	"github.com/berachain/beacon-kit/mod/primitives/pkg/crypto/sha256"
 	"github.com/berachain/beacon-kit/mod/primitives/pkg/math/log"
 	"github.com/berachain/beacon-kit/mod/primitives/pkg/math/pow"
 )
 
 type (
 	// GeneralizedIndex is a generalized index.
-	GeneralizedIndex[RootT ~[32]byte] uint64
+	GeneralizedIndex uint64
 
 	// GeneralizedIndices is a list of generalized indices.
-	GeneralizedIndices[RootT ~[32]byte] []GeneralizedIndex[RootT]
+	GeneralizedIndices []GeneralizedIndex
 )
 
 // NewGeneralizedIndex calculates the generalized index from the depth and
 // index.
-func NewGeneralizedIndex[RootT ~[32]byte](
+func NewGeneralizedIndex(
 	depth uint8,
 	index uint64,
-) GeneralizedIndex[RootT] {
-	return GeneralizedIndex[RootT]((1 << depth) + index)
+) GeneralizedIndex {
+	return GeneralizedIndex((1 << depth) + index)
 }
 
 // Unwrap returns the underlying uint64 value of the GeneralizedIndex.
-func (g GeneralizedIndex[RootT]) Unwrap() uint64 {
+func (g GeneralizedIndex) Unwrap() uint64 {
 	return uint64(g)
 }
 
 // Length returns the length of the generalized index.
-func (g GeneralizedIndex[RootT]) Length() int {
+func (g GeneralizedIndex) Length() int {
 	return int(log.ILog2Floor(g))
 }
 
 // IndexBit returns the bit at the specified position in a generalized index.
-func (g GeneralizedIndex[RootT]) IndexBit(position int) bool {
+func (g GeneralizedIndex) IndexBit(position int) bool {
 	return (g & (1 << position)) > 0
 }
 
 // Sibling returns the sibling index of the current generalized index.
-func (g GeneralizedIndex[RootT]) Sibling() GeneralizedIndex[RootT] {
+func (g GeneralizedIndex) Sibling() GeneralizedIndex {
 	return g ^ 1
 }
 
 // LeftChild returns the left child index of the current generalized index.
 //
 //nolint:mnd // from spec.
-func (g GeneralizedIndex[RootT]) LeftChild() GeneralizedIndex[RootT] {
+func (g GeneralizedIndex) LeftChild() GeneralizedIndex {
 	return 2 * g
 }
 
 // RightChild returns the right child index of the current generalized index.
-func (g GeneralizedIndex[RootT]) RightChild() GeneralizedIndex[RootT] {
+func (g GeneralizedIndex) RightChild() GeneralizedIndex {
 	return 2*g + 1
 }
 
 // Parent returns the parent index of the current generalized index.
 //
 //nolint:mnd // from spec.
-func (g GeneralizedIndex[RootT]) Parent() GeneralizedIndex[RootT] {
+func (g GeneralizedIndex) Parent() GeneralizedIndex {
 	return g / 2
 }
 
 // GetBranchIndices returns the generalized indices of the nodes on the path
 // from the root to the leaf.
-func (g GeneralizedIndex[RootT]) GetBranchIndices() GeneralizedIndices[RootT] {
+func (g GeneralizedIndex) GetBranchIndices() GeneralizedIndices {
 	// Get the generalized indices of the sister chunks along the path from the
 	// chunk with the given tree index to the root.
-	o := []GeneralizedIndex[RootT]{g.Sibling()}
+	o := GeneralizedIndices{g.Sibling()}
 	for o[len(o)-1] > 1 {
 		o = append(o, o[len(o)-1].Parent().Sibling())
 	}
@@ -99,51 +97,20 @@ func (g GeneralizedIndex[RootT]) GetBranchIndices() GeneralizedIndices[RootT] {
 
 // GetPathIndices returns the generalized indices of the nodes on the path from
 // the leaf to the root.
-func (g GeneralizedIndex[RootT]) GetPathIndices() GeneralizedIndices[RootT] {
+func (g GeneralizedIndex) GetPathIndices() GeneralizedIndices {
 	// Get the generalized indices of the sister chunks along the path from the
 	// chunk with the given tree index to the root.
-	o := []GeneralizedIndex[RootT]{g}
+	o := GeneralizedIndices{g}
 	for o[len(o)-1] > 1 {
 		o = append(o, o[len(o)-1].Parent())
 	}
 	return o[:len(o)-1]
 }
 
-// CalculateMerkleRoot calculates the Merkle root from the leaf and proof.
-func (g GeneralizedIndex[RootT]) CalculateMerkleRoot(
-	leaf RootT,
-	proof []RootT,
-) (RootT, error) {
-	if len(proof) != g.Length() {
-		return RootT{},
-			errors.Newf("expected proof length %d, received %d", g.Length(),
-				len(proof))
-	}
-	for i, h := range proof {
-		if g.IndexBit(i) {
-			leaf = sha256.Hash(append(h[:], leaf[:]...))
-		} else {
-			leaf = sha256.Hash(append(leaf[:], h[:]...))
-		}
-	}
-	return leaf, nil
-}
-
-// VerifyMerkleProof verifies the Merkle proof for the given
-// leaf, proof, and root.
-func (g GeneralizedIndex[RootT]) VerifyMerkleProof(
-	leaf RootT,
-	proof []RootT,
-	root RootT,
-) (bool, error) {
-	calculated, err := g.CalculateMerkleRoot(leaf, proof)
-	return calculated == root, err
-}
-
 // Concat multiple generalized indices into a single generalized index
 // representing the path from the first to the last node.
-func (gs GeneralizedIndices[RootT]) Concat() GeneralizedIndex[RootT] {
-	o := GeneralizedIndex[RootT](1)
+func (gs GeneralizedIndices) Concat() GeneralizedIndex {
+	o := GeneralizedIndex(1)
 	for _, i := range gs {
 		floorPower := pow.PrevPowerOfTwo(i)
 		o *= floorPower
@@ -156,11 +123,9 @@ func (gs GeneralizedIndices[RootT]) Concat() GeneralizedIndex[RootT] {
 // tree needed to prove the chunks with the given generalized indices. The
 // decreasing order is chosen deliberately to ensure equivalence to the order of
 // hashes in a regular single-item Merkle proof in the single-item case.
-func (
-	gs GeneralizedIndices[RootT],
-) GetHelperIndices() GeneralizedIndices[RootT] {
-	allHelperIndices := make(map[GeneralizedIndex[RootT]]struct{})
-	allPathIndices := make(map[GeneralizedIndex[RootT]]struct{})
+func (gs GeneralizedIndices) GetHelperIndices() GeneralizedIndices {
+	allHelperIndices := make(map[GeneralizedIndex]struct{})
+	allPathIndices := make(map[GeneralizedIndex]struct{})
 
 	for _, index := range gs {
 		for _, helperIndex := range index.GetBranchIndices() {
@@ -171,7 +136,7 @@ func (
 		}
 	}
 
-	difference := make([]GeneralizedIndex[RootT], 0, len(allHelperIndices))
+	difference := make(GeneralizedIndices, 0, len(allHelperIndices))
 	for helperIndex := range allHelperIndices {
 		if _, exists := allPathIndices[helperIndex]; !exists {
 			difference = append(difference, helperIndex)
@@ -179,80 +144,9 @@ func (
 	}
 
 	// Sort in decreasing order.
-	slices.SortFunc(difference, func(i, j GeneralizedIndex[RootT]) int {
+	slices.SortFunc(difference, func(i, j GeneralizedIndex) int {
 		return int(j - i)
 	})
 
 	return difference
-}
-
-// CalculateMultiMerkleRoot calculates the Merkle root for multiple leaves with
-// their corresponding proofs and indices.
-func (gs GeneralizedIndices[RootT]) CalculateMultiMerkleRoot(
-	leaves []RootT,
-	proof []RootT,
-) (RootT, error) {
-	if len(leaves) != len(gs) {
-		return RootT{}, errors.New(
-			"mismatched leaves and indices length",
-		)
-	}
-
-	helperIndices := gs.GetHelperIndices()
-	if len(proof) != len(helperIndices) {
-		return RootT{}, errors.New(
-			"mismatched proof and helper indices length",
-		)
-	}
-
-	objects := make(map[GeneralizedIndex[RootT]]RootT)
-	for i, index := range gs {
-		objects[index] = leaves[i]
-	}
-	for i, index := range helperIndices {
-		objects[index] = proof[i]
-	}
-
-	// Extract keys into slice to traverse in descending order.
-	keys := make([]GeneralizedIndex[RootT], 0, len(objects))
-	for k := range objects {
-		keys = append(keys, k)
-	}
-	slices.SortFunc(keys, func(i, j GeneralizedIndex[RootT]) int {
-		return int(j - i)
-	})
-
-	var (
-		pos     int
-		sibling RootT
-	)
-	for pos < len(keys) {
-		k := keys[pos]
-		if _, ok := objects[k]; ok {
-			if sibling, ok = objects[k^1]; ok {
-				if _, ok = objects[k/2]; !ok {
-					obj := objects[(k|1)^1]
-					objects[k/2] = sha256.Hash(append(obj[:], sibling[:]...))
-					//nolint:mnd // from spec.
-					keys = append(keys, k/2)
-				}
-			}
-		}
-		pos++
-	}
-	return objects[1], nil
-}
-
-// VerifyMerkleMultiproof verifies the Merkle multiproof by comparing the
-// calculated root with the provided root.
-func (gs GeneralizedIndices[RootT]) VerifyMerkleMultiproof(
-	leaves []RootT,
-	proof []RootT,
-	root RootT,
-) bool {
-	calculatedRoot, err := gs.CalculateMultiMerkleRoot(leaves, proof)
-	if err != nil {
-		return false
-	}
-	return calculatedRoot == root
 }
