+++ mod/primitives/pkg/encoding/ssz/merkle/tree.go
@@ -20,28 +20,115 @@
 
 package merkle
 
-import "github.com/berachain/beacon-kit/mod/primitives/pkg/math/pow"
+import (
+	"slices"
 
-// New returns a Merkle tree of the given leaves.
-// As defined in the Ethereum 2.0 Spec:
-// https://github.com/ethereum/consensus-specs/blob/dev/ssz/merkle-proofs.md#generalized-merkle-tree-index
-//
-//nolint:lll // link.
-func NewTree[LeafT ~[32]byte](
-	leaves []LeafT,
-	hashFn func([]byte) LeafT,
-) []LeafT {
-	/*
-	   Return an array representing the tree nodes by generalized index:
-	   [0, 1, 2, 3, 4, 5, 6, 7], where each layer is a power of 2. The 0 index is ignored. The 1 index is the root.
-	   The result will be twice the size as the padded bottom layer for the input leaves.
-	*/
-	bottomLength := pow.NextPowerOfTwo(uint64(len(leaves)))
-	//nolint:mnd // 2 is okay.
-	o := make([]LeafT, bottomLength*2)
-	copy(o[bottomLength:], leaves)
-	for i := bottomLength - 1; i > 0; i-- {
-		o[i] = hashFn(append(o[i*2][:], o[i*2+1][:]...))
+	"github.com/berachain/beacon-kit/mod/errors"
+	"github.com/berachain/beacon-kit/mod/primitives/pkg/crypto/sha256"
+)
+
+// CalculateMerkleRoot calculates the Merkle root from the leaf and proof.
+func CalculateMerkleRoot[RootT ~[32]byte](
+	index GeneralizedIndex,
+	leaf RootT,
+	proof []RootT,
+) (RootT, error) {
+	if len(proof) != index.Length() {
+		return RootT{},
+			errors.Newf("expected proof length %d, received %d", index.Length(),
+				len(proof))
 	}
-	return o
+	for i, h := range proof {
+		if index.IndexBit(i) {
+			leaf = sha256.Hash(append(h[:], leaf[:]...))
+		} else {
+			leaf = sha256.Hash(append(leaf[:], h[:]...))
+		}
+	}
+	return leaf, nil
+}
+
+// VerifyMerkleProof verifies the Merkle proof for the given
+// leaf, proof, and root.
+func VerifyMerkleProof[RootT ~[32]byte](
+	index GeneralizedIndex,
+	leaf RootT,
+	proof []RootT,
+	root RootT,
+) (bool, error) {
+	calculated, err := CalculateMerkleRoot(index, leaf, proof)
+	return calculated == root, err
+}
+
+// CalculateMultiMerkleRoot calculates the Merkle root for multiple leaves with
+// their corresponding proofs and indices.
+func CalculateMultiMerkleRoot[RootT ~[32]byte](
+	indices GeneralizedIndices,
+	leaves []RootT,
+	proof []RootT,
+) (RootT, error) {
+	if len(leaves) != len(indices) {
+		return RootT{}, errors.Newf(
+			"mismatched leaves and indices length: %d != %d",
+			len(leaves), len(indices),
+		)
+	}
+
+	helperIndices := indices.GetHelperIndices()
+	if len(proof) != len(helperIndices) {
+		return RootT{}, errors.New(
+			"mismatched proof and helper indices length",
+		)
+	}
+
+	objects := make(map[GeneralizedIndex]RootT)
+	for i, index := range indices {
+		objects[index] = leaves[i]
+	}
+	for i, index := range helperIndices {
+		objects[index] = proof[i]
+	}
+
+	// Extract keys into slice to traverse in descending order.
+	keys := make(GeneralizedIndices, 0, len(objects))
+	for k := range objects {
+		keys = append(keys, k)
+	}
+	slices.SortFunc(keys, func(i, j GeneralizedIndex) int {
+		return int(j - i)
+	})
+
+	var (
+		pos   int
+		left  RootT
+		right RootT
+	)
+	for pos < len(keys) {
+		k := keys[pos]
+		if _, ok := objects[k^1]; ok {
+			if _, ok = objects[k/2]; !ok {
+				left = objects[(k|1)^1]
+				right = objects[k|1]
+				objects[k/2] = sha256.Hash(append(left[:], right[:]...))
+				keys = append(keys, k/2) //nolint:mnd // from spec.
+			}
+		}
+		pos++
+	}
+	return objects[1], nil
+}
+
+// VerifyMerkleMultiproof verifies the Merkle multiproof by comparing the
+// calculated root with the provided root.
+func VerifyMerkleMultiproof[RootT ~[32]byte](
+	indices GeneralizedIndices,
+	leaves []RootT,
+	proof []RootT,
+	root RootT,
+) bool {
+	calculatedRoot, err := CalculateMultiMerkleRoot(indices, leaves, proof)
+	if err != nil {
+		return false
+	}
+	return calculatedRoot == root
 }
